refactor(publishing): share challenge abandon hook
